Adds notes to README.md explaining how to run site locally
This commit adds notes to the README.md file to explain how to run the site locally so that people can run the site on their machine. This is helpful when people need to edit it because it allows them to see full site while they edit. Another added benefit is that mkdocs has built-in hot reload which means whenever you make a change to a file and save it, the locally running site automatically refreshes.