Adds PDF artifact for offline viewing
Having a PDF version of the site is helpful for individuals that may want to
view an offline copy, replicate to another system (e.g. CIMUg SharePoint site),
or just prefer to read and browse the guide as a PDF. This commit adds a
new GitHub Action that, on release publish, will generate a PDF version of the
guide and publish it as an artifact.