Adds support for publishing multiple versions
This commit adds support for publishing multiple versions of the
modeling guide to the https://cim-mg.ucaiug.io site. It uses the "mike"
plugin to do this. See https://github.com/jimporter/mike and README.md
for details. It also removes the continuous integration GitHub Action so
that the site is no longer built and deployed automatically on push to
master since this results in "in-progress" versions of the guide being
available to the public before it is actually ready.