Adds enumeration plugin to automatmically create section numbers
Previously the modeling guide had manually entered section numbering.
Using the 'mkdocs-enumerate-headings-plugin' these section numbers can
be added automatically. This commit removes the manually entered ones
and replaces with the generated ones using the plugin. The changes also
ensure the generated PDF is being created with the same numbering scheme
as the site. Setting ordered_chapter_level to 0 for PDF generator plugin causes it to not generate any heading numbers and thus inherit the ones from the enumerate headings plugin.