Adds GitHub Action to publish https://cim-mg.ucaiug.io site
Adds GitHub Action that you can run manually as needed to publish version 1.1
of the https://cim-mg.ucaiug.io/ website. This isn't done autmomatically
on merge but rather only by manual/on-demand by navigating to the
"GitHub Action" tab of the repo and running. This is manual so that that the
CIM Modeling Guide maintainers have full control over when new versions
get published.