Adds GitHub Action to publish https://cim-mg.ucaiug.io site (#71)
Adds GitHub Action that you can to publish version 1.1 of the https://cim-mg.ucaiug.io
website. This isn't done autmomatically on merge but rather manually/on-demand by 
navigating to the "GitHub Action" tab of the repo and running it. This is manual so that
the CIM Modeling Guide maintainers have full control over when new versions get 
published.