Adds GitHub Action to publish https://cim-mg.ucaiug.io site (#72)
Adds GitHub Action so that you can publish versions of the https://cim-mg.ucaiug.io
website. This isn't done autmomatically on merge but rather manually/on-demand by
navigating to the "GitHub Action" tab of the repo and running it. This is manual so that
the CIM Modeling Guide maintainers have full control over when new versions get
published. Note that this action on master doesn't actually publish
because the "master" branch is considered in development and must not be
visible until a new version is ready. However, this GitHub Action is
needed on master so that users can still select it via the GitHub user
interface and so that it is ready once a new version (e.g. 2.0) needs to
be published.