Adds GitHub Action to publish v1.0 https://cim-mg.ucaiug.io site (#76)
Adds GitHub Action sothat you can publish version 1.0 of the https://cim-mg.ucaiug.io
website. This isn't done autmomatically on merge but rather manually/on-demand by
navigating to the "GitHub Action" tab of the repo and running it. This is manual so that
the CIM Modeling Guide maintainers have full control over when new versions get
published.