Adds gh-pages checkout to site publish workflow (#78)
The GitHub Action workflow for publishing the site was failing. This is
likely due to the gh-pages branch not getting checked out on GitHub Action
runs. It does not get checked out because by default the actions/checkout@v4
step does a very lightweight checkout (i.e. only current branch with depth of
1). This commit adds a fetch of the full gh-pages branch so that the branch is 
checked out and the publish step does not report missing remote changes.